Added return boolean for task eligibility
Return true when the task is runnable. False when the task is not
runnable. Note that this is different form whether the task has
succeeded or failed. (A task may be runnable and still fail.) If a
task returns false, that means we should probably try to run another,
different task instead.